fix: rename dynamic links to timeline links
