+++ frappe/core/doctype/communication/communication.py
@@ -8,10 +8,11 @@ from frappe.model.document import Document
 from frappe.utils import validate_email_address, get_fullname, strip_html, cstr
 from frappe.core.doctype.communication.email import (validate_email,
 	notify, _notify, update_parent_mins_to_first_response)
+from frappe.core.utils import get_parent_doc
 from frappe.utils.bot import BotReply
 from frappe.utils import parse_addr
 from frappe.core.doctype.comment.comment import update_comment_in_doc
-
+from email.utils import parseaddr
 from collections import Counter
 
 exclude_from_linked_with = True
@@ -56,9 +57,11 @@ class Communication(Document):
 		self.set_status()
 		self.set_sender_full_name()
 
-		self.validate_dynamic_links()
-		self.deduplicate_dynamic_links()
 		validate_email(self)
+
+		if self.communication_medium == "Email":
+			self.set_timeline_links()
+			self.deduplicate_timeline_links()
 
 	def validate_reference(self):
 		if self.reference_doctype and self.reference_name:
@@ -73,14 +76,13 @@ class Communication(Document):
 			# Prevent circular linking of Communication DocTypes
 			if self.reference_doctype == "Communication":
 				circular_linking = False
-				doc = get_parent_doc(self.reference_doctype, self.reference_name)
-				if doc:
-					while doc.reference_doctype == "Communication":
-						if doc:
-							if doc.reference_name == self.name:
-								circular_linking = True
-								break
-						doc = get_parent_doc(doc.reference_doctype, doc.reference_name)
+				doc = get_parent_doc(self)
+				while doc.reference_doctype == "Communication":
+					if get_parent_doc(doc).name==self.name:
+						circular_linking = True
+						break
+					doc = get_parent_doc(doc)
+
 				if circular_linking:
 					frappe.throw(_("Please make sure the Reference Communication Docs are not circularly linked."), frappe.CircularLinkingError)
 
@@ -234,11 +236,19 @@ class Communication(Document):
 				frappe.db.commit()
 
 	# Timeline Links
-	def deduplicate_dynamic_links(self):
-		if self.dynamic_links:
+	def set_timeline_links(self):
+		contacts = get_contacts([self.sender, self.recipients, self.cc, self.bcc])
+		for contact_name in contacts:
+			self.add_link('Contact', contact_name)
+
+			#link contact's dynamic links to communication
+			add_contact_links_to_communication(self, contact_name)
+
+	def deduplicate_timeline_links(self):
+		if self.timeline_links:
 			links, duplicate = [], False
 
-			for l in self.dynamic_links:
+			for l in self.timeline_links:
 				t = (l.link_doctype, l.link_name)
 				if not t in links:
 					links.append(t)
@@ -246,29 +256,12 @@ class Communication(Document):
 					duplicate = True
 
 			if duplicate:
-				del self.dynamic_links[:] # make it python 2 compatible as list.clear() is python 3 only
+				del self.timeline_links[:] # make it python 2 compatible as list.clear() is python 3 only
 				for l in links:
 					self.add_link(link_doctype=l[0], link_name=l[1])
 
-	def validate_dynamic_links(self):
-		circular_linking = False
-		for dynamic_link in self.dynamic_links:
-
-			# Prevent circular linking of Timeline DocTypes
-			if dynamic_link.link_doctype == "Communication":
-				doc = get_parent_doc(dynamic_link.link_doctype, dynamic_link.link_name)
-				if doc:
-					while doc.reference_doctype == "Communication":
-						if doc:
-							if doc.reference_name == self.name:
-								circular_linking = True
-								break
-
-		if circular_linking:
-			frappe.throw(_("Please make sure the Timeline Communication Docs are not circularly linked."), frappe.CircularLinkingError)
-
 	def add_link(self, link_doctype, link_name, autosave=False):
-		self.append("dynamic_links",
+		self.append("timeline_links",
 			{
 				"link_doctype": link_doctype,
 				"link_name": link_name
@@ -279,12 +272,12 @@ class Communication(Document):
 			self.save(ignore_permissions=True)
 
 	def get_links(self):
-		return self.dynamic_links
+		return self.timeline_links
 
 	def remove_link(self, link_doctype, link_name, autosave=False, ignore_permissions=True):
-		for l in self.dynamic_links:
+		for l in self.timeline_links:
 			if l.link_doctype == link_doctype and l.link_name == link_name:
-				self.dynamic_links.remove(l)
+				self.timeline_links.remove(l)
 
 		if autosave:
 			self.save(ignore_permissions=ignore_permissions)
@@ -292,7 +285,6 @@ class Communication(Document):
 def on_doctype_update():
 	"""Add indexes in `tabCommunication`"""
 	frappe.db.add_index("Communication", ["reference_doctype", "reference_name"])
-	frappe.db.add_index("Communication", ["link_doctype", "link_name"])
 	frappe.db.add_index("Communication", ["status", "communication_type"])
 
 def has_permission(doc, ptype, user):
@@ -323,8 +315,38 @@ def get_permission_query_conditions_for_communication(user):
 		return """tabCommunication.email_account in ({email_accounts})"""\
 			.format(email_accounts=','.join(email_accounts))
 
-def get_parent_doc(link_doctype, link_name):
-	"""Returns document of `link_doctype`, `link_name`"""
-	if link_doctype and link_name:
-		parent_doc = frappe.get_doc(link_doctype, link_name)
-	return parent_doc if parent_doc else None
+def get_contacts(email_strings):
+	email_addrs = []
+
+	for email_string in email_strings:
+		if email_string:
+			for email in email_string.split(","):
+					parsed_email = parseaddr(email)[1]
+					if parsed_email:
+						email_addrs.append(parsed_email)
+
+	contacts = []
+	for email in email_addrs:
+		contact_name = frappe.db.get_value('Contact', {'email_id': email})
+
+		if not contact_name:
+			contact = frappe.get_doc({
+					"doctype": "Contact",
+					"first_name": frappe.unscrub(email.split("@")[0]),
+					"email_id": email
+				}).insert(ignore_permissions=True)
+			contact_name = contact.name
+
+		contacts.append(contact_name)
+
+	return contacts
+
+def add_contact_links_to_communication(communication, contact_name):
+	contact_links = frappe.get_list("Dynamic Link", filters={
+			"parenttype": "Contact",
+			"parent": contact_name
+		}, fields=["link_doctype", "link_name"])
+
+	if contact_links:
+		for contact_link in contact_links:
+			communication.add_link(contact_link.link_doctype, contact_link.link_name)
